+++ components/Hamburger.js
@@ -1,11 +1,12 @@
-import React from 'react';
 import { FontAwesome5 } from '@expo/vector-icons';
+import React from 'react';
 import { HamburgerButton } from '../styled/hamburger';
 
 export default class Hamburger extends React.Component {
   constructor(props) {
     super(props);
   }
+
   render() {
     return (
       <HamburgerButton onPress={() => this.props.navigation.toggleDrawer()}>
