+++ lib/authUtils.js
@@ -1,5 +1,5 @@
+import Colors from '../constants/Colors';
 import BASE from './common';
-import Colors from '../assets/Colors';
 
 // Fields
