+++ screens/map/ProductsScreen.js
@@ -1,13 +1,13 @@
 import React from 'react';
-import { FlatList, View } from 'react-native';
+import { View } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
 import ProductCard from '../../components/product/ProductCard';
+import Colors from '../../constants/Colors';
 import {
-  ProductListHeaderContainer,
   ProductListContainer,
+  ProductListHeaderContainer,
   ProductListTitle,
 } from '../../styled/product';
-import Colors from '../../assets/Colors';
 
 class ProductsScreen extends React.Component {
   constructor(props) {
