+++ styled/rewards.js
@@ -1,7 +1,7 @@
 import { StyleSheet } from 'react-native';
 import styled from 'styled-components/native';
-import Colors from '../assets/Colors';
 import { NavButton } from '../components/BaseComponents';
+import Colors from '../constants/Colors';
 
 export const Container = styled.View`
   flex: 1;
